+++ DEVELOPMENT.md
@@ -1,2 +1,3 @@
 # Development
+- Run command locally: `dotnet run --project ./src/AustinS.TailwindCssTool -- [command]`
 - Pack for NuGet: `dotnet pack ./src/AustinS.TailwindCssTool -c Release /p:Version=X.Y.Z`
