+++ src/AustinS.TailwindCssTool/Program.cs
@@ -1,7 +1,8 @@
-﻿using AustinS.TailwindCssTool;
+﻿using AustinS.TailwindCssTool.Binary;
 using ConsoleAppFramework;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
+using BinaryManager = AustinS.TailwindCssTool.Binary.BinaryManager;
 
 var app = ConsoleApp
     .Create()
